Adding hand management and updating SDK to 2.3.b.210812

+++ UMI3D-Browser-Desktop/Assets/UMI3D SDK/ClientDevelopmentKit/InteractionSystem/Runtime/Loader/UMI3DInteractableLoader.cs
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+using inetum.unityUtils;
 using System;
 using umi3d.common;
 using umi3d.common.interaction;
@@ -47,7 +48,7 @@ namespace umi3d.cdk.interaction
                     break;
                 case UMI3DPropertyKeys.InteractableNodeId:
                     RemoveInteractableOnNode(dto);
-                    dto.nodeId = (ulong)property.value;
+                    dto.nodeId = (ulong)(long)property.value;
                     setInteractableOnNode(dto);
                     break;
                 case UMI3DPropertyKeys.InteractableHasPriority:
